+++ apps/zui/jest.config.js
@@ -15,7 +15,7 @@ module.exports = {
   setupFiles: ["./src/test/unit/setup/before-env.ts"],
   setupFilesAfterEnv: ["./src/test/unit/setup/after-env.ts"],
   testEnvironmentOptions: {
-    testURL: "http://localhost:3000/?name=search&id=test-1",
+    testURL: "http://localhost:4567/?name=search&id=test-1",
   },
   globalSetup: "./src/test/unit/setup/global.ts",
   modulePaths: ["<rootDir>"],
